+++ src/main/java/fpt/banking/system/service/TranferService.java
@@ -2,6 +2,6 @@ package fpt.banking.system.service;
 
 public interface TranferService {
 
-	public void tranferInternalByAccountNumber(int tranferAccountId, int receiveAccountId,
+	public void tranferInternalByAccountNumber(long tranferAccountId, long receiveAccountId,
 			long amount, String description);
 }
